src/api/pyproject.toml
[project]
name = "langchain-api"
authors = [
    {name = "Adrian Jimenez"}
]
description = "Langchain API"
readme = "README.md"
requires-python = ">=3.11"
version = "0.0.1"
dependencies = [
    "fastapi",
    "openai",
    "langchain",
    "langchain-openai",
    "langchain-experimental",
    "azure-identity",
    "azure-search-documents",
    "pyannote.audio",
    "cohere",
    "tenacity==8.3.0",
]

[project.optional-dependencies]
dev = [
    "python-dotenv",
    "ruff",
    "mypy",
]

[tool.setuptools]
package-dir = {"f_api" = "f_api"}

[tool.mypy]
python_version = "3.11"
strict = true
exclude = [
    '^setup\.py$',
    "venv",
]

[[tool.mypy.overrides]]
module = [
    "setuptools",
]
ignore_missing_imports = true

[tool.ruff]
select = [
    "E",    # pycodestyle
    "W",    # pycodestyle
    "F",    # pyflake
    "I",    # isort
    "S",    # flake8-bandit
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "D",    # pydocstyle
    "RUF",  # Ruff specific rules
]

exclude = [
    ".git",
    ".idea",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".tox",
    ".venv",
    "venv",
    "build",
    "assets",
    "docs/",
]

ignore = [
    "D100",     # Missing docstring in public module
    "D101",     # Missing docstring in public class
    "D102",     # Missing docstring in public method
    "D103",     # Missing docstring in public function
    "D107",     # Missing docstring in __init__
    "D203",     # no-blank-line-before-class (conflict with D211)
    "D213",     # multi-line-summary-second-line (conflict with D212)
    "S608",     # SQL Injection - Should be enabled at some point?
]

line-length = 120

[tool.ruff.per-file-ignores]
"__init__.py" = ["D104"]
"tests/*" = ["S", "D"]

[tool.ruff.isort]
combine-as-imports = true
force-wrap-aliases = true
